Make list options available from the player-options page
