+++ WebHostLib/static/styles/player-options.css
@@ -1,20 +1,3 @@
-html{
-    background-image: url('../static/backgrounds/grass.png');
-    background-repeat: repeat;
-    background-size: 650px 650px;
-}
-
-#player-options{
-    box-sizing: border-box;
-    max-width: 1024px;
-    margin-left: auto;
-    margin-right: auto;
-    background-color: rgba(0, 0, 0, 0.15);
-    border-radius: 8px;
-    padding: 1rem;
-    color: #eeffeb;
-}
-
 #player-options #player-options-button-row{
     display: flex;
     flex-direction: row;
@@ -22,84 +5,11 @@ html{
     margin-top: 15px;
 }
 
-#player-options code{
-    background-color: #d9cd8e;
-    border-radius: 4px;
-    padding-left: 0.25rem;
-    padding-right: 0.25rem;
-    color: #000000;
-}
-
-#player-options #user-message{
-    display: none;
-    width: calc(100% - 8px);
-    background-color: #ffe86b;
-    border-radius: 4px;
-    color: #000000;
-    padding: 4px;
-    text-align: center;
-}
-
-#player-options #user-message.visible{
-    display: block;
-    cursor: pointer;
-}
-
-#player-options h1{
-    font-size: 2.5rem;
-    font-weight: normal;
-    width: 100%;
-    margin-bottom: 0.5rem;
-    text-shadow: 1px 1px 4px #000000;
-}
-
-#player-options h2{
-    font-size: 40px;
-    font-weight: normal;
-    width: 100%;
-    margin-bottom: 0.5rem;
-    text-transform: lowercase;
-    text-shadow: 1px 1px 2px #000000;
-}
-
-#player-options h3, #player-options h4, #player-options h5, #player-options h6{
-    text-shadow: 1px 1px 1px rgba(0, 0, 0, 0.5);
-}
-
-#player-options input:not([type]){
-    border: 1px solid #000000;
-    padding: 3px;
-    border-radius: 3px;
-    min-width: 150px;
-}
-
-#player-options input:not([type]):focus{
-    border: 1px solid #ffffff;
-}
-
-#player-options select{
-    border: 1px solid #000000;
-    padding: 3px;
-    border-radius: 3px;
-    min-width: 150px;
-    background-color: #ffffff;
-}
-
-#player-options #game-options, #player-options #rom-options{
-    display: flex;
-    flex-direction: row;
-}
-
 #player-options #meta-options {
     display: flex;
     justify-content: space-between;
     gap: 20px;
     padding: 3px;
-}
-
-#player-options div {
-    display: flex;
-    flex-grow: 1;
 }
 
 #player-options #meta-options label {
@@ -116,19 +26,25 @@ html{
 }
 
 #player-options .left, #player-options .right{
-    flex-grow: 1;
+    width: 50%;
+    box-sizing: border-box;
 }
 
 #player-options .left{
-    margin-right: 10px;
+    float: left;
+    padding-right: 10px;
 }
 
 #player-options .right{
-    margin-left: 10px;
+    float: right;
+    padding-left: 10px;
+}
+
+#game-options-center {
+    clear: both;
 }
 
 #player-options table{
-    margin-bottom: 30px;
     width: 100%;
 }
 
@@ -208,6 +124,14 @@ html{
     vertical-align: top;
 }
 
+#player-options table .simple-list{
+    max-height: 120px;
+}
+
+#game-options-center{
+    margin-bottom: 30px;
+}
+
 @media all and (max-width: 1024px) {
     #player-options {
         border-radius: 0;
@@ -221,7 +145,6 @@ html{
 
     #player-options #game-options{
         justify-content: flex-start;
-        flex-wrap: wrap;
     }
 
     #player-options .left,
