+++ WebHostLib/static/styles/weighted-options.css
@@ -1,18 +1,5 @@
 html{
-    background-image: url('../static/backgrounds/grass.png');
-    background-repeat: repeat;
-    background-size: 650px 650px;
     scroll-padding-top: 90px;
-}
-
-#weighted-settings{
-    max-width: 1000px;
-    margin-left: auto;
-    margin-right: auto;
-    background-color: rgba(0, 0, 0, 0.15);
-    border-radius: 8px;
-    padding: 1rem;
-    color: #eeffeb;
 }
 
 #weighted-settings #games-wrapper{
@@ -42,15 +29,6 @@ html{
 
 #weighted-settings .setting-wrapper .add-option-div button:active{
     margin-bottom: 1px;
-}
-
-#weighted-settings p.setting-description{
-    margin: 0 0 1rem;
-}
-
-#weighted-settings p.hint-text{
-    margin: 0 0 1rem;
-    font-style: italic;
 }
 
 #weighted-settings .jump-link{
@@ -98,90 +76,6 @@ html{
     cursor: pointer;
 }
 
-#weighted-settings .items-wrapper{
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-}
-
-#weighted-settings .items-div h3{
-    margin-bottom: 0.5rem;
-}
-
-#weighted-settings .items-wrapper .item-set-wrapper{
-    width: 24%;
-    font-weight: bold;
-}
-
-#weighted-settings .item-container{
-    border: 1px solid #ffffff;
-    border-radius: 2px;
-    width: 100%;
-    height: 300px;
-    overflow-y: auto;
-    overflow-x: hidden;
-    margin-top: 0.125rem;
-    font-weight: normal;
-}
-
-#weighted-settings .item-container .item-div{
-    padding: 0.125rem 0.5rem;
-    cursor: pointer;
-}
-
-#weighted-settings .item-container .item-div:hover{
-    background-color: rgba(0, 0, 0, 0.1);
-}
-
-#weighted-settings .item-container .item-qty-div{
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    padding: 0.125rem 0.5rem;
-    cursor: pointer;
-}
-
-#weighted-settings .item-container .item-qty-div .item-qty-input-wrapper{
-    display: flex;
-    flex-direction: column;
-    justify-content: space-around;
-}
-
-#weighted-settings .item-container .item-qty-div input{
-    min-width: unset;
-    width: 1.5rem;
-    text-align: center;
-}
-
-#weighted-settings .item-container .item-qty-div:hover{
-    background-color: rgba(0, 0, 0, 0.1);
-}
-
-#weighted-settings .hints-div, #weighted-settings .locations-div{
-    margin-top: 2rem;
-}
-
-#weighted-settings .hints-div h3, #weighted-settings .locations-div h3{
-    margin-bottom: 0.5rem;
-}
-
-#weighted-settings .hints-container, #weighted-settings .locations-container{
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-}
-
-#weighted-settings .hints-wrapper, #weighted-settings .locations-wrapper{
-    width: calc(50% - 0.5rem);
-    font-weight: bold;
-}
-
-#weighted-settings .hints-wrapper .simple-list, #weighted-settings .locations-wrapper .simple-list{
-    margin-top: 0.25rem;
-    height: 300px;
-    font-weight: normal;
-}
-
 #weighted-settings #weighted-settings-button-row{
     display: flex;
     flex-direction: row;
@@ -189,113 +83,9 @@ html{
     margin-top: 15px;
 }
 
-#weighted-settings code{
-    background-color: #d9cd8e;
-    border-radius: 4px;
-    padding-left: 0.25rem;
-    padding-right: 0.25rem;
-    color: #000000;
-}
-
-#weighted-settings #user-message{
-    display: none;
-    width: calc(100% - 8px);
-    background-color: #ffe86b;
-    border-radius: 4px;
-    color: #000000;
-    padding: 4px;
-    text-align: center;
-}
-
-#weighted-settings #user-message.visible{
-    display: block;
-    cursor: pointer;
-}
-
-#weighted-settings h1{
-    font-size: 2.5rem;
-    font-weight: normal;
-    border-bottom: 1px solid #ffffff;
-    width: 100%;
-    margin-bottom: 0.5rem;
-    color: #ffffff;
-    text-shadow: 1px 1px 4px #000000;
-}
-
-#weighted-settings h2{
-    font-size: 2rem;
-    font-weight: normal;
-    border-bottom: 1px solid #ffffff;
-    width: 100%;
-    margin-bottom: 0.5rem;
-    color: #ffe993;
-    text-transform: none;
-    text-shadow: 1px 1px 2px #000000;
-}
-
-#weighted-settings h3, #weighted-settings h4, #weighted-settings h5, #weighted-settings h6{
-    color: #ffffff;
-    text-shadow: 1px 1px 1px rgba(0, 0, 0, 0.5);
-    text-transform: none;
-}
-
-#weighted-settings a{
-    color: #ffef00;
-    cursor: pointer;
-}
-
-#weighted-settings input:not([type]){
-    border: 1px solid #000000;
-    padding: 3px;
-    border-radius: 3px;
-    min-width: 150px;
-}
-
-#weighted-settings input:not([type]):focus{
-    border: 1px solid #ffffff;
-}
-
-#weighted-settings select{
-    border: 1px solid #000000;
-    padding: 3px;
-    border-radius: 3px;
-    min-width: 150px;
-    background-color: #ffffff;
-}
-
 #weighted-settings .game-options, #weighted-settings .rom-options{
     display: flex;
     flex-direction: column;
-}
-
-#weighted-settings .simple-list{
-    display: flex;
-    flex-direction: column;
-
-    max-height: 300px;
-    overflow-y: auto;
-    border: 1px solid #ffffff;
-    border-radius: 4px;
-}
-
-#weighted-settings .simple-list .list-row label{
-    display: block;
-    width: calc(100% - 0.5rem);
-    padding: 0.0625rem 0.25rem;
-}
-
-#weighted-settings .simple-list .list-row label:hover{
-    background-color: rgba(0, 0, 0, 0.1);
-}
-
-#weighted-settings .simple-list .list-row label input[type=checkbox]{
-    margin-right: 0.5rem;
-}
-
-#weighted-settings .simple-list hr{
-    width: calc(100% - 2px);
-    margin: 2px auto;
-    border-bottom: 1px solid rgb(255 255 255 / 0.6);
 }
 
 #weighted-settings .invisible{
